Add files via upload
This is a Gaussian Bayes classification model of the MNIST dataset. The business case it was being trained for was to be able to recognize numbers on doctor prescription forms. The text would then auto-populate into a system that the hospital is building as they are moving from paper-based information management to computer-based system.